+++ src/renderers/vulkan/inspector/vk_imgui.cpp
@@ -96,6 +96,7 @@ extern "C" void VkImgui_Initialize(void) {
 	if ( r_studio_tools && r_studio_tools->integer ) {
 		vkWindows.studioMap.open = qtrue;
 		vkWindows.studioConsole.open = qtrue;
+		vkWindows.studioAuthor.open = qtrue;
 	}
 
 	vkImguiState.active = qtrue;
@@ -187,6 +188,7 @@ extern "C" void VkImgui_Draw(void) {
 	VkImgui_DrawProfiler();
 	VkImgui_DrawStudioMapPanel();
 	VkImgui_DrawStudioConsolePanel();
+	VkImgui_DrawStudioAuthorPanel();
 
 	ImGui::End();
 	ImGui::Render();